+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.21] - 2025-07-14
+
+### Added
+- Save theme choice to local storage
+
+### Changed
+- Changed storybook background
+- Changed color on dark surface to be brighter
+
+### Fixed
+- Fixed `Menu` and `Overlay` z-Indexes
+
 ## [0.1.20] - 2025-07-14
 
 ### Added